requirements.txt
pytesseract
Pillow
playwright
pandas